\name{myRFMcalculation}
\alias{RFMfunction}
\title{Recency, Frequency, Monetary Value (RFM)}
\usage{
RFMfunction()
}
\description{
Calculate a weighted RFM score: recency, frequency, and monetary value for every customer
}
\usage{
RFMfunction(data, weight_recency = 1, weight_frequency = 1, weight_monetary = 1)
}
\examples{
hello()
}
