% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{RFMfunction}
\alias{RFMfunction}
\title{calculateRFM}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{data}{- A data.table containing the transaction recorded details for every customer.}

\item{r}{- Weight of recency}

\item{f}{- Weight of frequency}

\item{m}{- Weight of monetary value}
}
\description{
Description
  Calculate a weighted RFM score: recency, frequency, and monetary value for every customer
}
\details{
Arguments
}
